requirements.txt
numpy==1.19.5
tensorflow_gpu==2.6.0
